+++ Inventory-Management-System-With-DB/Models/Product.cs
@@ -6,9 +6,9 @@ namespace Inventory_Management_System.Models
     {
         public string Name { get; set; }
         public decimal Price { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
-        public Product(string name, decimal price, int quantity)
+        public Product(string name, decimal price, decimal quantity)
         {
             Name = name;
             Price = price;
